requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
pymupdf
python-docx
numpy
